Analyze a Large Dataset of Fraud in Financial Payment Services
For the first module of this project, we will need you to accomplish the following:

Set up your system by downloading the necessary files, tools and dependencies.
Explore the financial payment services dataset by completing the querying and data visualization exercises in the Additional Resources below
We've broken down the steps for you in stages below so you can accomplish this task in an organized manner.

Set Up:
Before you can tackle any software or development task you need to set up your development environment. Your development environment refers to your system having all the required software installed to modify the code, as well as getting the code of the project itself onto your computer.

To do this we've created a simple guide below on how to get your environment set up: Task1_setup.md

      2. Purpose:

Being able to query and visualize large datasets helps provide critical information for which areas need additional investment. This is not limited to financial systems but internal computer information systems as well.

     3. Acceptance Criteria:

Using the Pandas library, you learned how to explore and make sense of a dataset. In the process, you wrote queries and then visualized those queries in an easily digestible format that is presentable to management.